pyproject.toml
[tool.poetry]
name = "tbx"
version = "0.1.0"
description = "Torchbox.com"
authors = ["Torchbox Ltd"]

[tool.poetry.dependencies]
python = "^3.13"
Django = "~5.1"
wagtail = "~6.3"
psycopg = "^3.2.3"
gunicorn = "^23.0.0"
whitenoise = "^6.8.0"
phonenumbers = "^8.13.47"
Wand = "^0.6.10"

# Django packages
django-basic-auth-ip-whitelist = "^0.6"
django-birdbath = "^2.0.0"
django-csp = "^3.8"
django-pattern-library = "^1.3.0"
django-permissions-policy = "^4.22.0"
django-phonenumber-field = "^8.0.0"
django-redis = "^5.4.0"
django-referrer-policy = "~1.0"
django-storages = {version = "^1.14.4", extras = ["boto3"]}
dj-database-url = "^2.3.0"

# Wagtail packages
wagtail-accessibility = "^2.0.0"
wagtail-lite-youtube-embed = "^0.1.0"
wagtail-markdown = "^0.12.1"
wagtailmedia = "^0.15.2"
wagtail-purge = "^0.4.0"

# Production
scout-apm = "^3.2.1"
sentry-sdk = "^2.19.0"
html5lib = "^1.1"
beautifulsoup4 = "^4.12.3"
django-xff = "^1.4.0"

[tool.poetry.group.dev.dependencies]
Werkzeug = "^3.1.1"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
fabric = "^3.2.2"
pudb = "^2024.1"
honcho = "^2.0.0"

# Linters etc.
detect-secrets = "^1.5.0"
djhtml = "^3.0.6"
ruff = "^0.9.1"  # update pre-commit-config.yaml too
pre-commit = "^4.0.1"

# Documentation
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.41"
pymdown-extensions = "^10.11.2"

# Testing
factory-boy = "^3.3.0"
faker = "^33.3.0"
wagtail-factories = "^4.2.1"


[build-system]
requires = ["poetry>=1,<2"]
build-backend = "poetry.masonry.api"
